MGDAPI-4734 - move general functions outside of AWS package
