+++ pkg/providers/aws/provider_blobstorage.go
@@ -34,7 +34,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-//provider name and default create options
+// provider name and default create options
 const (
 	blobstorageProviderName               = "aws-s3"
 	defaultAwsBucketNameLength            = 40
@@ -109,7 +109,7 @@ func (p *BlobStorageProvider) GetReconcileTime(bs *v1alpha1.BlobStorage) time.Du
 	return resources.GetForcedReconcileTimeOrDefault(defaultReconcileTime)
 }
 
-//S3DeleteStrat custom s3 delete strat
+// S3DeleteStrat custom s3 delete strat
 type S3DeleteStrat struct {
 	_ struct{} `type:"structure"`
 
@@ -506,7 +506,7 @@ func (p *BlobStorageProvider) buildS3BucketConfig(ctx context.Context, bs *v1alp
 
 	// cluster infra info
 	p.Logger.Info("getting cluster id from infrastructure for bucket naming")
-	bucketName, err := BuildInfraNameFromObject(ctx, p.Client, bs.ObjectMeta, defaultAwsBucketNameLength)
+	bucketName, err := resources.BuildInfraNameFromObject(ctx, p.Client, bs.ObjectMeta, defaultAwsBucketNameLength)
 	if err != nil {
 		return nil, nil, nil, errorUtil.Wrapf(err, fmt.Sprintf("failed to retrieve aws s3 bucket config for blob storage instance %s", bs.Name))
 	}
@@ -569,7 +569,7 @@ func buildBlobStorageStatusMetricLabels(cr *v1alpha1.BlobStorage, clusterID, buc
 
 func (p *BlobStorageProvider) exposeBlobStorageMetrics(ctx context.Context, cr *v1alpha1.BlobStorage) {
 	// build instance name
-	bucketName, err := BuildInfraNameFromObject(ctx, p.Client, cr.ObjectMeta, defaultAwsBucketNameLength)
+	bucketName, err := resources.BuildInfraNameFromObject(ctx, p.Client, cr.ObjectMeta, defaultAwsBucketNameLength)
 	if err != nil {
 		logrus.Errorf("error occurred while building instance name during blob storage metrics: %v", err)
 	}
