+++ pkg/providers/aws/provider_postgres_metrics.go
@@ -4,13 +4,12 @@
 // alerts, to ensure and monitor performance of postgres (rds) instances
 //
 // this providers does
-//   * scrape metric data from cloudwatch
-//   * build a generic cloud metric data type from cloudwatch data
-//   * return generic cloud metric data to metric controller to be exposed
+//   - scrape metric data from cloudwatch
+//   - build a generic cloud metric data type from cloudwatch data
+//   - return generic cloud metric data to metric controller to be exposed
 //
 // this provider does not
-//   * expose the metrics, this is controller at a higher level (controller)
-//
+//   - expose the metrics, this is controller at a higher level (controller)
 package aws
 
 import (
@@ -100,7 +99,7 @@ func (p PostgresMetricsProvider) ScrapePostgresMetrics(ctx context.Context, post
 // scrapeRDSCloudWatchMetricData fetches cloud watch metrics for rds
 // and parses it to a GenericCloudMetric in order to return to the controller
 func (p *PostgresMetricsProvider) scrapeRDSCloudWatchMetricData(ctx context.Context, cloudWatchApi cloudwatchiface.CloudWatchAPI, postgres *v1alpha1.Postgres, metricTypes []providers.CloudProviderMetricType) ([]*providers.GenericCloudMetric, error) {
-	resourceID, err := BuildInfraNameFromObject(ctx, p.Client, postgres.ObjectMeta, defaultAwsIdentifierLength)
+	resourceID, err := resources.BuildInfraNameFromObject(ctx, p.Client, postgres.ObjectMeta, defaultAwsIdentifierLength)
 	if err != nil {
 		return nil, errorUtil.Errorf("error occurred building instance name: %v", err)
 	}
